refactor: apply clippy lints, rustfmt, and idiomatic Rust patterns
Systematic cleanup across existing modules to improve code quality
without changing behavior:

- actor.rs: Reorder init_test() after function definitions. Replace
  panic!() with assert!() in test assertions. Remove unnecessary
  .clone() calls. Apply rustfmt formatting fixes.

- session.rs: Use Self instead of Endpoint in impl blocks. Add
  #[must_use] to channel() method. Combine redundant match arms.
  Move init_test() after function definitions. Apply rustfmt.

- remote.rs: Use finish_non_exhaustive() for Debug impls. Replace
  match with map_or for option handling. Combine error match arms
  with shared handling. Apply rustfmt formatting.

- lab/network/harness.rs: Add doc comments to struct fields. Add
  #[must_use] annotations. Change pass-by-value to pass-by-reference
  for CancelRequest, LeaseRenewal, RemoteMessage parameters. Use
  u128::from() instead of as-cast. Use finish_non_exhaustive().

- lab/network/config.rs: Use mul_add for LogNormal computation.

- lab/network/network.rs: Use u128::from() for bytes_to_nanos.

- trace/canonicalize.rs: Use Vec::len method reference. Change
  kind_discriminant to take by value. Use qualified TraceEventKind::
  variants. Use u64::from() and u64::try_from() instead of as-casts.
  Combine match arms.

- trace/independence.rs: Combine match arms for IO events, cancel
  ack, and region operations. Use qualified TraceEventKind variants.
  Apply rustfmt formatting.

- trace/distributed/lattice.rs: Combine match arms, simplify patterns.

- trace/distributed/sheaf.rs: Add #[must_use], apply rustfmt.

- trace/distributed/mod.rs: Rustfmt import ordering.

+++ src/lab/network/network.rs
@@ -481,7 +481,7 @@ fn bytes_to_nanos(len: usize, bandwidth: u64) -> u64 {
     if len == 0 || bandwidth == 0 {
         return 0;
     }
-    let nanos = (len as u128)
+    let nanos = u128::from(len as u64)
         .saturating_mul(1_000_000_000u128)
         .saturating_div(u128::from(bandwidth));
     nanos.min(u128::from(u64::MAX)) as u64
